Refactor reducers
